Refactored error handling components.
