+++ src/cache/provider.tsx
@@ -11,58 +11,58 @@ type Payload = {
   };
 };
 
+function getCurrentHours() {
+  const secondsInHours = 3600;
+
+  const currentHours = new Date().getTime() / secondsInHours;
+
+  return currentHours;
+}
+
 function set<Data>(key: Keys, user: string, data: Data, limit: number): void {
-  try {
-    const secondsInHours = 3600;
+  const current = getCurrentHours();
 
-    const current = new Date().getTime() / secondsInHours;
+  const time = { limit, current };
 
-    const time = { limit, current };
+  const payload: Payload = { user, data, time };
 
-    const payload: Payload = { user, data, time };
+  const string = JSON.stringify(payload);
 
-    const string = JSON.stringify(payload);
-
-    localStorage.setItem(key, string);
-  } catch (error) {}
+  localStorage.setItem(key, string);
 }
 
 function get<Data, Revert>(
   key: Keys,
   user: string,
   revert: Revert
-): Data | Revert | undefined {
-  try {
-    const secondsInHours = 3600;
+): Data | Revert {
+  const cache = localStorage.getItem(key);
 
-    const cache = localStorage.getItem(key);
+  const isCacheNull = cache === null;
 
-    const isCacheNull = cache === null;
-
-    if (isCacheNull) {
-      return revert;
-    }
-
-    const payload: Payload = JSON.parse(cache);
-
-    const isUserValid = user === payload.user;
-
-    if (!isUserValid) {
-      return revert;
-    }
-
-    const updated = new Date().getTime() / secondsInHours;
-
-    const differace = updated - payload.time.current;
-
-    const isValid = differace < payload.time.limit;
-
-    if (isValid) {
-      return payload.data as Data;
-    }
-
+  if (isCacheNull) {
     return revert;
-  } catch (error) {}
+  }
+
+  const payload: Payload = JSON.parse(cache);
+
+  const isUserValid = user === payload.user;
+
+  if (!isUserValid) {
+    return revert;
+  }
+
+  const updatedHours = getCurrentHours();
+
+  const differaceInHours = updatedHours - payload.time.current;
+
+  const isValidDifference = differaceInHours < payload.time.limit;
+
+  if (isValidDifference) {
+    return payload.data as Data;
+  }
+
+  return revert;
 }
 
 const clientCache = { set, get };
